+++ src/commands/EditorController.ts
@@ -4,6 +4,35 @@ import GitService from '../utils/GitService';
 import EditorView from '../webviews/EditorView';
 import UiApi from '../utils/UiApi';
 
+type ViewColumnKey =
+  | 'Active'
+  | 'Beside'
+  | 'One'
+  | 'Two'
+  | 'Three'
+  | 'Four'
+  | 'Five'
+  | 'Six'
+  | 'Seven'
+  | 'Eight'
+  | 'Nine';
+
+type ViewColumnMap = Record<ViewColumnKey, number>;
+
+const editorGroupMap: ViewColumnMap = {
+  Active: vscode.ViewColumn.Active,
+  Beside: vscode.ViewColumn.Beside,
+  One: vscode.ViewColumn.One,
+  Two: vscode.ViewColumn.Two,
+  Three: vscode.ViewColumn.Three,
+  Four: vscode.ViewColumn.Four,
+  Five: vscode.ViewColumn.Five,
+  Six: vscode.ViewColumn.Six,
+  Seven: vscode.ViewColumn.Seven,
+  Eight: vscode.ViewColumn.Eight,
+  Nine: vscode.ViewColumn.Nine,
+};
+
 export default class EditorController {
   private _primaryEditorPanel: vscode.WebviewPanel | undefined;
   private _ui: UiApi | undefined;
@@ -14,13 +43,15 @@ export default class EditorController {
   ) {}
 
   openInTheMainView() {
+    const config = vscode.workspace.getConfiguration('commit-message-editor');
+
     if (this._primaryEditorPanel) {
       this._primaryEditorPanel.reveal();
     } else {
       this._primaryEditorPanel = vscode.window.createWebviewPanel(
         ViewType.PrimaryEditor,
         'Commit Message Editor',
-        vscode.ViewColumn.One,
+        editorGroupMap[config.view.columnToShowIn as ViewColumnKey],
         {
           enableScripts: true,
           retainContextWhenHidden: true,